Update 2000-01-05-what-is-a-transaction.md
I was wondering if the transaction could be stated as: "the process that takes place when value is transferred between entities on the blockchain."

I was thinking that when Bob wants to send funds to anyone else, from the perspective of the pending transaction, the prior transactions don't matter.  All that matters is that (Bob's funds available) >= (funds to be sent + transaction fee). Do I have that wrong? The blockchain, though, WILL have a record of the prior funds received by Bob that go towards the funds he's sending.